requirements.txt
Django==2.2
djangorestframework==3.9.2
django-cors-headers==2.1.0
django-storages==1.7.1
django-extensions==2.1.9
moto==1.3.8
psycopg2==2.7.3
django-crispy-forms==1.7.2
pysaml2==4.5.0
uwsgi==2.0.17
xlwt==1.3.0
requests==2.22.0
pyjwt==1.7.1
